+++ src/refinegems/__init__.py
@@ -1,8 +1,8 @@
-__all__ = ['analysis','classes','curation','decorators','utility', 'cmd_access']
+__all__ = ['analysis','classes','curation','developement','utility', 'cmd_access']
 
-# from . import analysis
-# from . import classes
-# from . import curation
-# from . import decorators
-# from . import utility
-# from . import cmd_access
+from . import analysis
+from . import classes
+from . import curation
+from . import developement
+from . import utility
+from . import cmd_access
